Fixed methodological error
Disability weights error fix
----------------------------
Previously, interventions reduced disability on the `combined_dw` state.
This state is a blend of healthy disability weights, and copd disability
weights.
Therefore, the interventions were incorrectly reducing disability on the
combined state, which would overcount their effect.
We have rerouted this logic, so only the COPD disability weight is
affected, before it is combined.